refactor/store to separate files
