--- src/store/films.json
@@ -1,90 +1,0 @@
-[
-  {
-    "id": "1",
-    "title": "Pulp Fiction",
-    "url": "pulp_fiction.png",
-    "director": "Quentin Tarantino",
-    "year": 1994,
-    "duration": 154,
-    "rating": 4.3,
-    "description": "Hitmen Jules Winnfield and Vincent Vega arrive at an apartment to retrieve a briefcase for their boss, gangster Marsellus Wallace, from a business partner, Brett. After Vincent checks the contents of the briefcase, Jules shoots one of Brett's associates. He declaims a passage from the Bible, and he and Vincent kill Brett for trying to double-cross Marsellus. They take the briefcase to Marsellus and wait while he bribes boxer Butch Coolidge to take a dive in his upcoming match",
-    "genre": ["Action", "Adventure"]
-  },
-  {
-    "id": "2",
-    "title": "Four rooms",
-    "url": "four_rooms.png",
-    "director": "Allison Anders",
-    "year": 1995,
-    "duration": 172,
-    "rating": 4.5,
-    "description": "This movie features the collaborative directorial efforts of four new filmmakers, each of whom directs a segment of this comedy. It's New Year's Eve at the Mon Signor Hotel, a former grand old Hollywood hotel, now fallen upon hard times. Often using physical comedy and sight gags, this movie chronicles the slapstick misadventures of Ted, the Bellhop. He's on his first night on the job, when he's asked to help out a coven of witches in the Honeymoon Suite.",
-    "genre": ["Comedy", "Anthology"]
-  },
-  {
-    "id": "3",
-    "title": "Avengers",
-    "url": "avengers.png",
-    "director": "Joss Whedon",
-    "year": 2012,
-    "duration": 142,
-    "rating": 4.6,
-    "description": "The Asgardian Loki encounters the Other, the leader of an extraterrestrial race known as the Chitauri. In exchange for retrieving the Tesseract,a powerful energy source of unknown potential, the Other promises Loki an army with which he can subjugate Earth.",
-    "genre": ["Comedy", "Fantasy"]
-  },
-  {
-    "id": "4",
-    "title": "Bonemian Rhapsody",
-    "url": "bonemian.png",
-    "director": "Bryan Singer",
-    "year": 2018,
-    "duration": 134,
-    "rating": 4.3,
-    "description": "In 1985, Queen waits to perform at Live Aid. Fifteen years earlier, in 1970, Farrokh Bulsara works as a baggage handler at Heathrow Airport and lives with his parents and sister. He goes to a pub to see the band Smile perform. Seeking them out after the show, he meets and is attracted to Mary Austin, who works at the boutique Biba. ",
-    "genre": ["Musical"]
-  },
-  {
-    "id": "5",
-    "title": "Inception",
-    "url": "inception.png",
-    "director": "Christopher Nolan",
-    "year": 2010,
-    "duration": 148,
-    "rating": 4.7,
-    "description": "Cobb and Arthur are they perform corporate espionage using experimental dream-sharing technology to infiltrate their targets' subconscious and extract information. Their latest target, Saito, is impressed with Cobb's ability to layer multiple dreams within each other and offers to hire Cobb for the supposedly impossible job of implanting an idea into a person's subconscious; performing 'inception' on Robert, the son of Saito's competitor Maurice Fischer,",
-    "genre": ["block Baster"]
-  },
-  {
-    "id": "6",
-    "title": "Jackie Brown",
-    "url": "jackie_brown.png",
-    "director": "Quentin Tarantino",
-    "year": 1997,
-    "duration": 154,
-    "rating": 4.8,
-    "description": "Jackie Brown, a flight attendant, smuggles money from Mexico into the United States for Ordell Robbie, a gun runner in Los Angeles. When Ordell's courier, Beaumont Livingston, is arrested, he hires bail bondsman Max Cherry to bail him out. To prevent Beaumont from talking to the police, Ordell kills him.",
-    "genre": ["block Baster"]
-  },
-  {
-    "id": "7",
-    "title": "Kill Bill",
-    "url": "kill_bill.png",
-    "director": "Quentin Tarantino",
-    "year": 2003,
-    "duration": 111,
-    "rating": 4.5,
-    "description": "In 1999, a pregnant woman in a wedding dress, the Bride, lies wounded in a chapel in El Paso, Texas. She tells her attacker, Bill, that the baby is his just as he shoots her in the head. Years later, the Bride, having survived the attack, goes to the home of Vernita Green, planning to kill her. Both women were members of the Deadly Vipers, a now-disbanded group of assassins led by Bill. Vernita now leads a normal suburban family life.",
-    "genre": ["block Baster"]
-  },
-  {
-    "id": "8",
-    "title": "Reservoir Dogs",
-    "url": "reservoir_dogs.png",
-    "director": "Quentin Tarantino",
-    "year": 1992,
-    "duration": 99,
-    "rating": 4.5,
-    "description": "Eight gangsters eat breakfast at a Los Angeles diner. All but the boss Joe Cabot and his son, underboss 'Nice Guy' Eddie Cabot, use aliases: Mr. Brown, Mr. White, Mr. Blonde, Mr. Blue, Mr. Orange and Mr. Pink. After Mr. Brown finishes rambling about the Madonna song 'Like a Virgin', the group argues about Mr. Pink's policy of not tipping.",
-    "genre": ["block Baster"]
-  }
-]
